+++ NEWS.md
@@ -1,3 +1,9 @@
+# hlatools 0.1.12
+
+* "etnicity" replaced with "ancestry" in IPD-IMGT/HLA XML source
+* Accessor function "etnicity()" replaced with "ancestry()".
+* Update for IPD-IMGT/HLA up to 3.53.0
+
 # hlatools 0.1.11
 
 * Update for IPD-IMGT/HLA 3.48.0, 3.49.0, and 3.50.0